third_party/blink/web_tests/fast/multicol/client-rects-empty-element-boundary.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>body { margin:0; }</style>
<div style="columns:4; column-gap:10px; column-fill:auto; width:430px; height:100px;">
    <div style="height:100px;"></div>
    <!-- A zero-block-size element fits at the bottom of a fragmentainer. -->
    <div id="empty1" style="position:relative;"></div>
    <div style="height:100px;">
      <div id="empty2" style="position:relative;"></div>
    </div>
</div>
<script>
    test(function() {
        let rects = empty1.getClientRects();
        assert_equals(rects.length, 1);
        assert_equals(rects[0].left, 0);
        assert_equals(rects[0].top, 100);
        assert_equals(rects[0].right, 100);
        assert_equals(rects[0].bottom, 100);

        rects = empty2.getClientRects();
        assert_equals(rects.length, 1);
        assert_equals(rects[0].left, 110);
        assert_equals(rects[0].top, 0);
        assert_equals(rects[0].right, 210);
        assert_equals(rects[0].bottom, 0);
    }, "Zero-height element at column boundary");
</script>
